Move lat/lon parsing code out of waypoint code, which is getting long.
